destroy virsh

+++ destroy_vms.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 if [[ $(command -v VBoxManage) != '' ]]
 then
-	while true 
+	while true
 	do
 		VBoxManage list runningvms | grep shutit_openshift_cluster | awk '{print $1}' | xargs -IXXX VBoxManage controlvm 'XXX' poweroff && VBoxManage list vms | grep shutit_openshift_cluster | awk '{print $1}'  | xargs -IXXX VBoxManage unregistervm 'XXX' --delete
 		# The xargs removes whitespace
@@ -14,7 +14,9 @@ then
 		fi
 	done
 fi
-if [[ $(kvm-ok 2>&1 | command grep 'can be used') != '' ]]
-then                                                                                                                                                                              
-    virsh list | grep shutit_openshift_cluster | awk '{print $1}' | xargs -n1 virsh destroy                                                                                   
-fi                                                                                              
+if [[ $(command -v virsh) != '' ]]
+	if [[ $(kvm-ok 2>&1 | command grep 'can be used') != '' ]]
+	then
+	    virsh list | grep shutit_openshift_cluster | awk '{print $1}' | xargs -n1 virsh destroy
+	fi
+fi
